feat: improve error messaging when picking csv stock file
