--- src/main/java/edu/ntnu/idatt2003/group14/exception/csvReading/CSVError.java
@@ -1,8 +1,0 @@
-package edu.ntnu.idatt2003.group14.exception.csvReading;
-
-public enum CSVError {
-  FILE_NOT_FOUND,
-  READ_FAILED,
-  EMPTY_FILE,
-  PARSING
-}
