--- src/main/java/edu/ntnu/idatt2003/group14/exception/csvReading/ParsingError.java
@@ -1,7 +1,0 @@
-package edu.ntnu.idatt2003.group14.exception.csvReading;
-
-public record ParsingError(
-    int errorLineNumber,
-    String errorLineString
-) {
-}
