+++ src/main/java/edu/ntnu/idatt2003/group14/ui/features/menu/newgame/NewGameController.java
@@ -1,18 +1,18 @@
 package edu.ntnu.idatt2003.group14.ui.features.menu.newgame;
 
-import edu.ntnu.idatt2003.group14.exception.csvReading.CSVReadException;
+import edu.ntnu.idatt2003.group14.exception.csvreading.CSVReadException;
 import edu.ntnu.idatt2003.group14.logging.AppLogger;
 import edu.ntnu.idatt2003.group14.service.GameService;
 import edu.ntnu.idatt2003.group14.ui.app.AppRouter;
 import edu.ntnu.idatt2003.group14.ui.app.Route;
 import java.io.File;
-import java.io.IOException;
 import java.math.BigDecimal;
 
 /**
  * Controller class for handling user input from the New Game scene.
  *
  * @author Elias Haugsbakk, Kevin Holswilder
+ * @version 1.0.0
  * @since 0.0.1
  */
 public class NewGameController {
@@ -44,7 +44,7 @@ public class NewGameController {
    * @param amount        the amount of starting money
    * @param stockDataFile the stock data {@link File}
    * @return the {@link NewGameValidationState} containing information
-   * on if the inputted information is valid
+   *      on if the inputted information is valid
    */
   public NewGameValidationState validateNewGameInput(String username, String amount,
                                                      File stockDataFile) {
