+++ src/main/java/edu/ntnu/idatt2003/group14/ui/features/menu/newgame/NewGameView.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idatt2003.group14.ui.features.menu.newgame;
 
-import edu.ntnu.idatt2003.group14.exception.csvReading.CSVReadException;
+import edu.ntnu.idatt2003.group14.exception.csvreading.CSVReadException;
 import edu.ntnu.idatt2003.group14.service.AudioManager;
 import edu.ntnu.idatt2003.group14.ui.app.AppController;
 import edu.ntnu.idatt2003.group14.ui.app.View;
@@ -26,6 +26,7 @@ import javafx.scene.layout.VBox;
  * and stock file and lets them start the game.</p>
  *
  * @author Elias Haugsbakk
+ * @version 1.0.0
  * @since 0.0.1
  */
 public class NewGameView implements View {
@@ -153,7 +154,6 @@ public class NewGameView implements View {
               this.stockDataFile
           );
         } catch (CSVReadException e) {
-          IO.println("debugger parsing error: " + e.getError().name());
           switch (e.getError()) {
             case READ_FAILED ->
                 showError("The reading of the CSV file failed; see millions.log", fileChooserBtn);
@@ -161,11 +161,11 @@ public class NewGameView implements View {
                 showError("CSV file not found; see millions.log", fileChooserBtn);
             case EMPTY_FILE ->
                 showError("The CSV file does not contain any stocks", fileChooserBtn);
-            case PARSING -> {
-              showError(
-                  "Could not pares line " + e.getParsingError().errorLineNumber() + ": " + "\"" +
-                      e.getParsingError().errorLineString() + "\"", fileChooserBtn);
-            }
+            case PARSING -> showError(
+                "Could not parse line " + e.getParsingError().errorLineNumber() + ": " + "\""
+                    + e.getParsingError().errorLineString() + "\"", fileChooserBtn);
+            default -> showError("Unexpected error when parsing CSV file; view millions.log",
+                fileChooserBtn);
           }
         }
       }
@@ -182,10 +182,13 @@ public class NewGameView implements View {
           "Username cannot be empty",
           usernameField
       );
-      case NO_FILE_CHOSEN -> showError(
-          "No stock data file has been selected",
-          fileChooserBtn
-      );
+      case NO_FILE_CHOSEN -> {
+        showError(
+            "No stock data file has been selected",
+            fileChooserBtn
+        );
+        this.fileChooserBtn.setText("Pick CSV stock data file");
+      }
       default -> throw new IllegalStateException("Unexpected startingMoney: " + result);
     }
   }
